Move input related source files into src/input/
