+++ src/input/meson.build
@@ -1,0 +1,6 @@
+labwc_sources += files(
+  'keyboard.c',
+  'key-state.c',
+  'touch.c',
+  'cursor.c',
+)
